rename __version__.py to version.py
