+++ opencensus/stats/exporters/stackdriver_exporter.py
@@ -23,11 +23,11 @@ from datetime import datetime
 from google.api_core.gapic_v1 import client_info
 from google.cloud import monitoring_v3
 
-from opencensus.common.__version__ import __version__
 from opencensus.common import utils
 from opencensus.common.monitored_resource_util.monitored_resource_util \
     import MonitoredResourceUtil
 from opencensus.common.transports import async_
+from opencensus.common.version import __version__
 from opencensus.stats import aggregation
 from opencensus.stats import measure
 from opencensus.stats.exporters import base
