+++ tests/unit/trace/exporters/ocagent/test_trace_exporter.py
@@ -19,7 +19,7 @@ import os
 import socket
 import unittest
 
-from opencensus.common.__version__ import __version__
+from opencensus.common.version import __version__
 from opencensus.trace import span_context as span_context_module
 from opencensus.trace import span_data as span_data_module
 from opencensus.trace.exporters.gen.opencensus.trace.v1 import trace_config_pb2
